Refactor zoom components to have general reusability
