+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/components/ScatterPlot/index.js
@@ -1,0 +1,1 @@
+export { default } from './ScatterPlot'
