+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/helpers/utils.js
@@ -1,5 +1,12 @@
 import { scaleLinear } from '@vx/scale'
 
+export function getDataExtent(data) {
+  return {
+    x: d3.extent(data.x),
+    y: d3.extent(data.y)
+  }
+}
+
 export function left (tickDirection, margin) {
   const left = {
     inner: 0,
@@ -18,7 +25,7 @@ export function top (tickDirection, margin) {
   return top[tickDirection]
 }
 
-export function xMin (tickDirection, padding) {
+export function xMin ({ tickDirection, padding }) {
   const xMin = {
     inner: padding.left,
     outer: 0
@@ -27,7 +34,7 @@ export function xMin (tickDirection, padding) {
   return xMin[tickDirection]
 }
 
-export function xMax (tickDirection, parentWidth, margin) {
+export function xMax ({ tickDirection, parentWidth, margin }) {
   const xMax = {
     inner: parentWidth - margin.left,
     outer: parentWidth - margin.left - margin.right
@@ -36,7 +43,7 @@ export function xMax (tickDirection, parentWidth, margin) {
   return xMax[tickDirection]
 }
 
-export function yMin (tickDirection, padding) {
+export function yMin ({ tickDirection, padding }) {
   const yMin = {
     inner: padding.bottom,
     outer: 0
@@ -45,7 +52,7 @@ export function yMin (tickDirection, padding) {
   return yMin[tickDirection]
 }
 
-export function yMax (tickDirection, parentHeight, margin, padding) {
+export function yMax ({ tickDirection, parentHeight, margin, padding }) {
  const yMax = {
    inner: parentHeight - padding.bottom,
    outer: parentHeight - margin.top - margin.bottom
@@ -54,31 +61,37 @@ export function yMax (tickDirection, parentHeight, margin, padding) {
  return yMax[tickDirection]
 }
 
-export function scaleTransform (dataExtent, transformMatrix, xRange, yRange) {
+export function transformXScale (data, transformMatrix, rangeParameters) {
+  const dataExtent = getDataExtent(data)
+  const xRange = [xMin(rangeParameters), xMax(rangeParameters)]
   const xScale = scaleLinear({
     domain: dataExtent.x,
     range: xRange
   })
+
+  return scaleLinear({
+    domain: [
+      xScale.invert((xScale(dataExtent.x[0]) - transformMatrix.translateX) / transformMatrix.scaleX),
+      xScale.invert((xScale(dataExtent.x[1]) - transformMatrix.translateX) / transformMatrix.scaleX)
+    ],
+    range: xRange
+  })
+} 
+
+export function transformYScale (data, transformMatrix, rangeParameters) {
+  const dataExtent = getDataExtent(data)
+  const yRange = [yMin(rangeParameters, yMax(rangeParameters))]
 
   const yScale = scaleLinear({
     domain: dataExtent.y,
     range: yRange
   })
 
-  return {
-    xScale: scaleLinear({
-      domain: [
-        xScale.invert((xScale(dataExtent.x[0]) - transformMatrix.translateX) / transformMatrix.scaleX),
-        xScale.invert((xScale(dataExtent.x[1]) - transformMatrix.translateX) / transformMatrix.scaleX)
-      ],
-      range: xRange
-    }),
-    yScale: scaleLinear({
-      domain: [
-        yScale.invert((yScale(dataExtent.y[0]) - transformMatrix.translateY) / transformMatrix.scaleY),
-        yScale.invert((yScale(dataExtent.y[1]) - transformMatrix.translateY) / transformMatrix.scaleY)
-      ],
-      range: yRange,
-    })
-  }
+  return scaleLinear({
+    domain: [
+      yScale.invert((yScale(dataExtent.y[0]) - transformMatrix.translateY) / transformMatrix.scaleY),
+      yScale.invert((yScale(dataExtent.y[1]) - transformMatrix.translateY) / transformMatrix.scaleY)
+    ],
+    range: yRange,
+  })
 }
